refactor: notification-service Module generate and refactor
