+++ notification-service/src/main/java/com/whalewatch/domain/AlertSetting.java
