+++ notification-service/src/main/java/com/whalewatch/domain/UserAlert.java
