+++ notification-service/src/main/java/com/whalewatch/dto/AlertSettingsDto.java
