+++ notification-service/src/main/java/com/whalewatch/dto/UserAlertDto.java
