+++ notification-service/src/main/java/com/whalewatch/redis/RedisStateService.java
@@ -1,4 +1,4 @@
-package com.whalewatch.service;
+package com.whalewatch.redis;
 
 import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.stereotype.Service;
