+++ notification-service/src/main/java/com/whalewatch/service/UserAlertService.java
