+++ service/build.gradle
@@ -18,7 +18,6 @@ dependencies {
 
     //Security
     implementation 'org.springframework.boot:spring-boot-starter-security'
-
     testImplementation 'org.springframework.boot:spring-boot-starter-test'
 
     //telegramBot
